% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_prep.R
\name{df_prep}
\alias{df_prep}
\title{df_prep}
\usage{
df_prep(data, id, activity, time, mood, max.time)
}
\arguments{
\item{data}{Data frame. Experience sampling/day reconstruction data that includes variables collecting each individual ID, current mood, current activity, time and any additional variables that the users might want to include in their analyses.}

\item{id}{String. Name of participant id variable in data.}

\item{activity}{String. Name of activity variable in data.}

\item{time}{String. Name of time stamp variable in data.Variable must be of POSIXct class.}

\item{mood}{String. Name of mood variable in data.}

\item{max.time}{Maximum time allowed between two consecutive mood observations (in hours).}
}
\description{
Prepares raw data to run mood homeostasis estimation.
}
\keyword{Affect}
\keyword{Dynamics}
\keyword{Homeostasis,}
\keyword{Mood}
