% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mood_homeostasis.R
\name{mood_homeostasis}
\alias{mood_homeostasis}
\title{mood_homeostasis}
\usage{
mood_homeostasis(data, activity, global = TRUE)
}
\arguments{
\item{data}{Data frame. Experience sampling/day reconstruction data that includes variables collecting each individual ID, current mood, current activity, time and any additional variables that the users might want to include in their analyses.}

\item{activity}{String. Name of activity variable in data.}

\item{global}{Whether global estimates of the effect of activity on mood are used in the computation.If TRUE, the effect fo each activity on mood is estimated using the pooled data. If FALSE, an individual-specific estimate of the effect of each activity on mood is estimated and used in the computation.}
}
\description{
Uses the data frame obtained from df_prep to obtain estimates of mood homeostasis for each participant.
}
\keyword{Affect}
\keyword{Dynamics}
\keyword{Homeostasis,}
\keyword{Mood}
